+++ src/test/java/ibm/jceplus/junit/openjceplusfips/TestAll.java
@@ -47,8 +47,6 @@ import org.junit.platform.suite.api.Suite;
     TestHmacSHA224InteropSunJCE.class,
     TestHmacSHA256.class,
     TestHmacSHA256InteropSunJCE.class,
-    TestHmacSHA3_224.class,
-    TestHmacSHA3_256.class,
     TestHmacSHA3_384.class,
     TestHmacSHA3_512.class,
     TestHmacSHA384.class,
